funciona con ts, muy pocas modificaciones hechas
