+++ app/controllers/getUsers.ts
@@ -1,7 +1,5 @@
-const getUsers = (req, res) => {
+const getUsers = (req: any, res: any): void => {
   res.send([{ name: "David", age: 37, pass: 123 }, `Request url: ${req.url}`])
 }
 
-module.exports = {
-  getUsers,
-}
+export default getUsers
