--- app/controllers/postTime.js
@@ -1,6 +1,0 @@
-const postTime = (req, res) => {
-  const day = new Date()
-  res.send(day)
-}
-
-module.exports = postTime
