+++ app/controllers/postTime.ts
@@ -1,0 +1,6 @@
+const postTime = (_req: any, res: any) => {
+  const day = new Date()
+  res.send(day)
+}
+
+export default postTime
