+++ app/middlewares/nivell2exercici1.ts
@@ -1,6 +1,6 @@
-const express = require("express")
 
-const middle = (req, res, next) => {
+
+const middle = (req: any, res: any, next: any) => {
   res.set("Cache-Control", "no-cache")
   res.header("Access-Control-Allow-Origin", "*")
   req.body = { name: "David", age: 37, password: "asd" }
@@ -12,4 +12,4 @@ const middle = (req, res, next) => {
   next()
 }
 
-module.exports = middle
+export default middle
