--- app/routes/imgRouter.js
@@ -1,8 +1,0 @@
-const express = require("express")
-const { postImg } = require("../controllers/postImg")
-
-const router = express.Router()
-
-router.post("/img", postImg)
-
-module.exports = router
