+++ app/routes/imgRouter.ts
@@ -1,0 +1,8 @@
+import express from "express"
+import postImg from "../controllers/postImg"
+
+const router = express.Router()
+
+router.post("/img", postImg)
+
+export default router
