+++ app/routes/pokemonRouter.ts
@@ -1,0 +1,8 @@
+import express from 'express'
+import getPokemon from "../controllers/getPokemon"
+
+const router = express.Router()
+
+router.get("/pokemon/:id", getPokemon)
+
+export default router
