--- app/routes/timeRouter.js
@@ -1,8 +1,0 @@
-const express = require("express")
-const postTime = require("../controllers/postTime")
-
-const router = express.Router()
-
-router.post("/time", postTime)
-
-module.exports = router
