--- app/routes/users.js
@@ -1,8 +1,0 @@
-const express = require("express")
-const { getUsers } = require("../controllers/getUsers")
-
-const router = express.Router()
-
-router.get("/", getUsers)
-
-module.exports = router
