+++ app/routes/users.ts
@@ -1,0 +1,9 @@
+import express from 'express'
+import getUsers from '../controllers/getUsers'
+
+
+const router = express.Router()
+
+router.get("/", getUsers)
+
+export default router
