+++ notas_typescript.md
@@ -1,0 +1,4 @@
+tsc -init ---> crea el archivo tsconfig.json
+
+    * "rootDir": -----> donde se encuentran nuestros archivos de typescript
+    * "outDir": -----> destino de la compilacion
