+++ package.json
@@ -4,8 +4,7 @@
   "description": "sample project for node",
   "main": "app.js",
   "scripts": {
-    "start": "node app/app.js",
-    "dev": "nodemon app/app.js",
+    "start": "nodemon build/app.js",
     "entrega42": "jest"
   },
   "author": "",
@@ -22,9 +21,12 @@
     "path": "^0.12.7",
     "sequelize": "^6.29.1",
     "sequelize-cli": "^6.6.0",
-    "sqlite3": "^5.1.4"
+    "sqlite3": "^5.1.4",
+    "typescript": "^5.0.3"
   },
   "devDependencies": {
+    "@types/express": "^4.17.17",
+    "@types/node": "^18.15.11",
     "jest": "^28.1.1",
     "nodemon": "^2.0.7"
   }
